退运费区分查询与执行校验

+++ jmp/mp-service/src/main/java/com/vpu/mp/service/pojo/wxapp/order/OrderBeforeParam.java
@@ -51,8 +51,8 @@ public class OrderBeforeParam extends AbstractOrderOperateQueryParam{
      * 0:默认选第一张；null：不选；其他：卡号
      */
     private String couponSn;
-    /**积分抵扣金额*/
-    private BigDecimal scoreDiscount;
+    /**积分抵扣*/
+    private Integer scoreDiscount;
     /**余额抵扣金额*/
     private BigDecimal balance;
     /**会员卡抵扣金额*/
